refactor: use the right file name convention
