+++ packages/agent/src/routes/modification/action/errors/approval-not-allowed-error.ts
